+++ crates/fosr-lib/src/inject/fast.rs
@@ -1,6 +1,7 @@
 use super::NetEnabler;
 use crate::stats::Stats;
 use crate::structs::*;
+
 use crossbeam_channel::Receiver;
 use crossbeam_channel::RecvTimeoutError;
 use crossbeam_channel::TryRecvError;