+++ crates/fosr/tests/cli_tests.rs
@@ -34,7 +34,7 @@ fn deterministic_fast_generation() -> Result<(), Box<dyn std::error::Error>> {
     let hash = sha256.finalize();
     assert_eq!(
         hex::encode(hash),
-        "131529091e2b4330dbc24f65c85b04657587cd6258ed1b1320516a27b54c60b4"
+        "633e1a7cc866c23778f6fa3ef6d5501c5cb4e0007409a937dc62ec821caa5626"
     );
     Ok(())
 }
@@ -65,7 +65,7 @@ fn deterministic_efficient_generation() -> Result<(), Box<dyn std::error::Error>
     let hash = sha256.finalize();
     assert_eq!(
         hex::encode(hash),
-        "131529091e2b4330dbc24f65c85b04657587cd6258ed1b1320516a27b54c60b4"
+        "633e1a7cc866c23778f6fa3ef6d5501c5cb4e0007409a937dc62ec821caa5626"
     );
     Ok(())
 }